pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pmk_probes"]

[project]
name = "pmk-probes"
version = "1.0.9"
authors = [
  { name="Felix Wenk", email="felix.wenk@pmk.de" },
]
description = "A Python package to control PMK's active oscilloscope probes."
readme = "README.md"
requires-python = ">=3.10"
license = "BSD-3-Clause"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial >= 3.5",
]

[project.urls]
Homepage = "http://www.pmk.de/"
Documentation = "https://pmk-probes.readthedocs.io/en/latest/"
Repository = "https://github.com/PMK-Probing-Solutions/pmk-probes"
